refactor: SkeletonBlock, CommonLoading을 common/loading 폴더로 이동 및 storybook 추가
