+++ src/components/main/plan/route/SkeletonScheduleRoutesContent.tsx
@@ -1,4 +1,4 @@
-import SkeletonBlock from "@/components/common/SkeletonBlock";
+import SkeletonBlock from "@/components/common/loading/SkeletonBlock";
 
 /**
  * ScheduleRoutesContent 스켈레톤
